+++ packages/sit-onyx/src/components/OnyxBadge/types.ts
@@ -5,7 +5,7 @@ export type OnyxBadgeProps = DensityProp & {
   /**
    * The color of the badge.
    */
-  variation?: OnyxColor;
+  color?: OnyxColor;
   /**
    * An icon which will be displayed inside the badge.
    * Note: If the icon property is set, the default slot will not be rendered.